hygiene, clarify how to compare counts and number of target items in rrul_fill_*() routines

+++ src/evrrul.c
@@ -860,7 +860,7 @@ rrul_fill_yly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	bool ymdp;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1021,7 +1021,7 @@ rrul_fill_mly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	bool ymdp;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1151,7 +1151,7 @@ rrul_fill_wly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	uint_fast32_t wd_incs = 0UL;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1295,7 +1295,7 @@ rrul_fill_dly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	unsigned int maxd;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1435,7 +1435,7 @@ rrul_fill_Hly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	uint_fast32_t H_mask = 0U;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1612,7 +1612,7 @@ rrul_fill_Mly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	uint_fast64_t M_mask = 0U;
 	struct enum_s e;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
@@ -1792,7 +1792,7 @@ rrul_fill_Sly(echs_instant_t *restrict tgt, size_t nti, rrulsp_t rr)
 	uint_fast64_t M_mask = 0U;
 	uint_fast64_t S_mask = 0U;
 
-	if (UNLIKELY(rr->count < nti)) {
+	if (UNLIKELY((unsigned int)rr->count < nti)) {
 		if (UNLIKELY((nti = rr->count) == 0UL)) {
 			goto fin;
 		}
